Dementia_API/requirements.txt
fastapi
uvicorn
transformers
torch
librosa
soundfile
python-multipart
google-genai
python-dotenv